requirements.txt

streamlit==1.37.1
pandas==2.2.2
pdfplumber==0.11.4
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
